--- l10n_it_account_balance_report/README.rst
@@ -1,98 +1,0 @@
-==========================================
-ITA - Stato patrimoniale e conto economico
-==========================================
-
-.. !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-   !! This file is generated by oca-gen-addon-readme !!
-   !! changes will be overwritten.                   !!
-   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-
-.. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
-    :target: https://odoo-community.org/page/development-status
-    :alt: Beta
-.. |badge2| image:: https://img.shields.io/badge/licence-AGPL--3-blue.png
-    :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
-    :alt: License: AGPL-3
-.. |badge3| image:: https://img.shields.io/badge/github-OCA%2Fl10n--italy-lightgray.png?logo=github
-    :target: https://github.com/OCA/l10n-italy/tree/14.0/l10n_it_account_balance_report
-    :alt: OCA/l10n-italy
-.. |badge4| image:: https://img.shields.io/badge/weblate-Translate%20me-F47D42.png
-    :target: https://translation.odoo-community.org/projects/l10n-italy-14-0/l10n-italy-14-0-l10n_it_account_balance_report
-    :alt: Translate me on Weblate
-.. |badge5| image:: https://img.shields.io/badge/runbot-Try%20me-875A7B.png
-    :target: https://runbot.odoo-community.org/runbot/122/14.0
-    :alt: Try me on Runbot
-
-|badge1| |badge2| |badge3| |badge4| |badge5| 
-
-**Italiano**
-
-Rendicontazione PDF e XLS per stato patrimoniale e conto economico a sezioni contrapposte.
-
-**English**
-
-PDF and XLS reporting for accounting balance sheets.
-
-**Table of contents**
-
-.. contents::
-   :local:
-
-Usage
-=====
-
-From your Accounting / Report section, select "Account Balance Report".
-
-This will trigger a wizard, from which you can setup your report configuration.
-From it, you can choose whether should be created, either an HTML interactive view, or a PDF / XLS sheet.
-
-Bug Tracker
-===========
-
-Bugs are tracked on `GitHub Issues <https://github.com/OCA/l10n-italy/issues>`_.
-In case of trouble, please check there if your issue has already been reported.
-If you spotted it first, help us smashing it by providing a detailed and welcomed
-`feedback <https://github.com/OCA/l10n-italy/issues/new?body=module:%20l10n_it_account_balance_report%0Aversion:%2014.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
-
-Do not contact contributors directly about support or help with technical issues.
-
-Credits
-=======
-
-Authors
-~~~~~~~
-
-* Openforce
-
-Contributors
-~~~~~~~~~~~~
-
-* Silvio Gregorini <https://github.com/SilvioGregorini>
-* `Ooops <https://www.ooops404.com>`_
-* `TAKOBI <https://takobi.online>`_:
-
-  * Simone Rubino <sir@takobi.online>
-
-Other credits
-~~~~~~~~~~~~~
-
-The development of this module has been financially supported by:
-
-* Odoo Italia Network
-
-Maintainers
-~~~~~~~~~~~
-
-This module is maintained by the OCA.
-
-.. image:: https://odoo-community.org/logo.png
-   :alt: Odoo Community Association
-   :target: https://odoo-community.org
-
-OCA, or the Odoo Community Association, is a nonprofit organization whose
-mission is to support the collaborative development of Odoo features and
-promote its widespread use.
-
-This module is part of the `OCA/l10n-italy <https://github.com/OCA/l10n-italy/tree/14.0/l10n_it_account_balance_report>`_ project on GitHub.
-
-You are welcome to contribute. To learn how please visit https://odoo-community.org/page/Contribute.
