--- l10n_it_account_balance_report/__init__.py
@@ -1,5 +1,0 @@
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-
-from . import models
-from . import report
-from . import wizard
