--- l10n_it_account_balance_report/__manifest__.py
@@ -1,30 +1,0 @@
-# Copyright 2019 Openforce (http://www.openforce.it)
-# Copyright 2019 Alessandro Camilli (alessandrocamilli@openforce.it)
-# Copyright 2019 Silvio Gregorini (silviogregorini@openforce.it)
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-
-{
-    "name": "ITA - Stato patrimoniale e conto economico",
-    "summary": "Rendicontazione .pdf e .xls per stato patrimoniale e conto"
-    " economico a sezioni contrapposte",
-    "version": "14.0.1.0.0",
-    "category": "Localization/Italy",
-    "author": "Odoo Community Association (OCA), Openforce",
-    "website": "https://github.com/OCA/l10n-italy"
-    "/tree/14.0/l10n_it_account_balance_report",
-    "license": "AGPL-3",
-    "depends": [
-        "account_financial_report",
-        "account_menu",
-        "l10n_it_account",
-    ],
-    "data": [
-        "data/account_type.xml",
-        "report/templates/account_balance_report.xml",
-        "report/reports.xml",
-        "views/account_types.xml",
-        "views/assets.xml",
-        "wizard/wizard_account_balance_report.xml",
-    ],
-    "installable": True,
-}
