--- l10n_it_account_balance_report/readme/DESCRIPTION.rst
@@ -1,7 +1,0 @@
-**Italiano**
-
-Rendicontazione PDF e XLS per stato patrimoniale e conto economico a sezioni contrapposte.
-
-**English**
-
-PDF and XLS reporting for accounting balance sheets.
