--- l10n_it_account_balance_report/readme/USAGE.rst
@@ -1,4 +1,0 @@
-From your Accounting / Report section, select "Account Balance Report".
-
-This will trigger a wizard, from which you can setup your report configuration.
-From it, you can choose whether should be created, either an HTML interactive view, or a PDF / XLS sheet.
