--- l10n_it_account_balance_report/report/__init__.py
@@ -1,4 +1,0 @@
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-
-from . import account_balance_report
-from . import account_balance_report_xlsx
