--- l10n_it_account_balance_report/report/reports.xml
@@ -1,40 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<odoo>
-
-    <record id="report_account_balance_report_html" model="ir.actions.report">
-        <field name="model">trial.balance.report.wizard</field>
-        <field name="name">Account Balance Report HTML</field>
-        <field name="report_type">qweb-html</field>
-        <field
-            name="report_name"
-        >l10n_it_account_balance_report.account_balance_report</field>
-        <field
-            name="report_file"
-        >l10n_it_account_balance_report.account_balance_report</field>
-    </record>
-
-    <record id="report_account_balance_report_pdf" model="ir.actions.report">
-        <field name="model">trial.balance.report.wizard</field>
-        <field name="name">Account Balance Report PDF</field>
-        <field name="report_type">qweb-pdf</field>
-        <field
-            name="paperformat_id"
-            ref="l10n_it_account.l10n_it_account_a4_landscape"
-        />
-        <field
-            name="report_name"
-        >l10n_it_account_balance_report.account_balance_report</field>
-        <field
-            name="report_file"
-        >l10n_it_account_balance_report.account_balance_report</field>
-    </record>
-
-    <record id="report_account_balance_report_xlsx" model="ir.actions.report">
-        <field name="model">trial.balance.report.wizard</field>
-        <field name="name">Account Balance Report XLSX</field>
-        <field name="report_type">xlsx</field>
-        <field name="report_name">l10n_it_account_balance_report.report_xlsx</field>
-        <field name="report_file">l10n_it_account_balance_report.report_xlsx</field>
-    </record>
-
-</odoo>
