--- l10n_it_account_balance_report/tests/__init__.py
@@ -1,3 +1,0 @@
-#  License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
-
-from . import test_balance_report
