--- l10n_it_account_balance_report/wizard/__init__.py
@@ -1,3 +1,0 @@
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-
-from . import wizard_account_balance_report
