+++ l10n_it_financial_statements_report/__init__.py
@@ -1,0 +1,6 @@
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+
+from . import models
+from . import report
+from . import wizard
+from .hooks import pre_absorb_old_module
