+++ l10n_it_financial_statements_report/readme/CONTRIBUTORS.rst
