+++ l10n_it_financial_statements_report/readme/DESCRIPTION.rst
@@ -1,0 +1,7 @@
+**Italiano**
+
+Rendicontazione PDF e XLS per stato patrimoniale e conto economico a sezioni contrapposte.
+
+**English**
+
+PDF and XLS reporting for financial statements.
