+++ l10n_it_financial_statements_report/readme/INSTALL.rst
@@ -1,0 +1,7 @@
+**Italiano**
+
+Se questo modulo viene installato in un ambiente dove `l10n_it_account_balance_report` è presente, lo sostituisce.
+
+**English**
+
+If this module is installed in an instance where `l10n_it_account_balance_report` is present, it replaces it.
