+++ l10n_it_financial_statements_report/readme/USAGE.rst
@@ -1,0 +1,4 @@
+From your Accounting / Report section, select "Financial Statements Report".
+
+This will trigger a wizard, from which you can setup your report configuration.
+From it, you can choose whether should be created, either an HTML interactive view, or a PDF / XLS sheet.
