+++ l10n_it_financial_statements_report/report/__init__.py
@@ -1,0 +1,4 @@
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+
+from . import financial_statements_report
+from . import financial_statements_report_xlsx
