+++ l10n_it_financial_statements_report/tests/__init__.py
@@ -1,0 +1,3 @@
+#  License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
+
+from . import test_report
