+++ l10n_it_financial_statements_report/wizard/__init__.py
@@ -1,0 +1,3 @@
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+
+from . import wizard_financial_statements_report
