--- setup/l10n_it_account_balance_report/odoo/addons/l10n_it_account_balance_report
@@ -1,1 +1,0 @@
-../../../../l10n_it_account_balance_report
