+++ setup/l10n_it_financial_statements_report/odoo/addons/l10n_it_financial_statements_report
@@ -1,0 +1,1 @@
+../../../../l10n_it_financial_statements_report
